fix: isolate studio runtime source loading
