+++ packages/cli/src/server/studioServer.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import { loadHyperframeRuntimeSource } from "@hyperframes/core";
-import { loadRuntimeSourceFallback } from "./studioServer.js";
+import { loadRuntimeSourceFallback } from "./runtimeSource.js";
 
 describe("loadRuntimeSourceFallback", () => {
   it("loads runtime source from the published core entrypoint", async () => {
